+++ src/examples/google_chatbot/api.py
@@ -1,60 +1,27 @@
-from typing import List, Optional
-
-from steamship import Block
-from steamship.agents.base import Metadata
-from steamship.agents.context.context import AgentContext
-from steamship.agents.llm.openai import OpenAI
-from steamship.agents.planner.react import ReACTPlanner
 from steamship.agents.service.agent_service import AgentService
 from steamship.agents.tools.image_generation.google_image_search import GoogleImageSearchTool
 from steamship.agents.tools.search.search import SearchTool
 from steamship.utils.repl import AgentREPL
 
-from core_prompt_builder import make_core_prompt
+from core.agent import BaseAgent
+from core.personality import Personality
 
 
-class GoogleChatbot(AgentService):
+class GoogleChatbot(BaseAgent):
     """Deployable Multimodal Agent that lets you talk to Google Search & Google Images.
 
     NOTE: To extend and deploy this agent, copy and paste the code into api.py.
 
     """
 
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        # The agent's planner is responsible for making decisions about what to do for a given input.
-        self.planner = ReACTPlanner(
-            tools=[
-                SearchTool(),
-                GoogleImageSearchTool()
-            ],
-            llm=OpenAI(self.client),
-        )
-        self.planner.PROMPT = self.get_personality()
+    def get_tools(self):
+        return [
+            SearchTool(),
+            GoogleImageSearchTool()
+        ]
 
-    def create_response(self, context: AgentContext) -> Optional[List[Block]]:
-
-        if len(context.emit_funcs) == 0:
-            context.emit_funcs.append(self._send_message_agent)
-
-        if len(context.chat_history.messages) == 0:
-            context.chat_history.append_system_message()
-
-        self.run_agent(context)
-
-        # should we return any message to the user to indicate that a response?
-        # maybe: "Working on it..." or "Received: {prompt}..."
-        return []
-
-    def _send_message_agent(self, blocks: List[Block], meta: Metadata):
-        # should this be directly-referenced, or should this be an invoke() endpoint, with a value passed
-        # in?
-
-        print(f"\n\nTELEGRAM SENDING MESSAGES:\n{blocks}")
-        # self.telegram_transport.send(messages)
-
-    def get_personality(self):
-        return make_core_prompt(
+    def get_personality(self) -> Personality:
+        return Personality(
             name="Assistant",
             byline="an assistant who helps search the web.",
             identity=[
